Create book.cpp
Developed a C++ program to create an efficient inventory management system for a bookshop.Designed and implemented a class called Books with appropriate member functions and constructors to store and manage book details, such as author, title, price, publisher, and stock position. Utilized the new operator in constructors to dynamically allocate memory space required for the book data.